backend/requirements.txt
Flask==2.2.2
Werkzeug==2.3.8
python-dotenv==0.21.0
requests==2.28.1
stability-sdk==0.2.2
Flask-Cors==3.0.10
apify-client==1.2.1
